Refactoring methods from QueryProvider into mixin classes
